+++ crates/vala/vala-bifrost-redux/src/tables/metrics/points.rs
@@ -20,13 +20,13 @@ use wyrd_spec::vala::api::PhysicalLayoutWire;
 use wyrd_spec::vala::managed_columns::WYRD_EVENT_TIME;
 
 /// Element declaration of an explicit-bucket count collection.
-pub static BUCKET_COUNT_ELEMENT: F = F::payload(19, "item", T::UInt64, false);
+pub static BUCKET_COUNT_ELEMENT: F = F::payload(19, "item", T::Int64, false);
 
 /// Element declaration of an explicit-bound collection.
 pub static EXPLICIT_BOUND_ELEMENT: F = F::payload(21, "item", T::Float64, false);
 
 /// Element declaration of the positive exponential bucket counts.
-pub static POSITIVE_BUCKET_COUNT_ELEMENT: F = F::payload(28, "item", T::UInt64, false);
+pub static POSITIVE_BUCKET_COUNT_ELEMENT: F = F::payload(28, "item", T::Int64, false);
 
 /// Ordered fields of the positive exponential bucket collection.
 pub static POSITIVE_BUCKET_FIELDS: [F; 2] = [
@@ -40,7 +40,7 @@ pub static POSITIVE_BUCKET_FIELDS: [F; 2] = [
 ];
 
 /// Element declaration of the negative exponential bucket counts.
-pub static NEGATIVE_BUCKET_COUNT_ELEMENT: F = F::payload(32, "item", T::UInt64, false);
+pub static NEGATIVE_BUCKET_COUNT_ELEMENT: F = F::payload(32, "item", T::Int64, false);
 
 /// Ordered fields of the negative exponential bucket collection.
 pub static NEGATIVE_BUCKET_FIELDS: [F; 2] = [
@@ -72,7 +72,7 @@ pub static QUANTILE_VALUE_ELEMENT: F = F::payload(
 /// An exemplar carries filtered caller attributes and the correlation ids of
 /// the originating span, so the whole subtree is access-gated payload.
 pub static EXEMPLAR_FIELDS: [F; 6] = [
-    F::sensitive(41, "time_unix_nano", T::UInt64, false),
+    F::sensitive(41, "time_unix_nano", T::Int64, false),
     F::sensitive(42, "int_value", T::Int64, true),
     F::sensitive(43, "double_value", T::Float64, true),
     F::sensitive(44, "filtered_attributes", T::Binary, false),
@@ -98,15 +98,15 @@ pub static METRIC_FIELDS: &[CanonicalField] = &[
     F::meta(3, "unit", T::Utf8, false),
     F::sensitive(4, "metadata", T::Binary, false),
     F::meta(5, "metric_type", T::Utf8, false),
-    F::meta(6, "time_unix_nano", T::UInt64, false),
-    F::meta(7, "start_time_unix_nano", T::UInt64, false),
-    F::meta(8, "flags", T::UInt32, false),
+    F::meta(6, "time_unix_nano", T::Int64, false),
+    F::meta(7, "start_time_unix_nano", T::Int64, false),
+    F::meta(8, "flags", T::Int64, false),
     F::sensitive(9, "attributes", T::Binary, false),
     F::meta(10, "int_value", T::Int64, true),
     F::meta(11, "double_value", T::Float64, true),
     F::meta(12, "aggregation_temporality", T::Int32, true),
     F::meta(13, "is_monotonic", T::Bool, true),
-    F::meta(14, "histogram_count", T::UInt64, true),
+    F::meta(14, "histogram_count", T::Int64, true),
     F::meta(15, "histogram_sum", T::Float64, true),
     F::meta(16, "histogram_min", T::Float64, true),
     F::meta(17, "histogram_max", T::Float64, true),
@@ -118,7 +118,7 @@ pub static METRIC_FIELDS: &[CanonicalField] = &[
         true,
     ),
     F::meta(22, "exponential_scale", T::Int32, true),
-    F::meta(23, "exponential_zero_count", T::UInt64, true),
+    F::meta(23, "exponential_zero_count", T::Int64, true),
     F::meta(24, "exponential_zero_threshold", T::Float64, true),
     F::payload(
         25,
@@ -132,7 +132,7 @@ pub static METRIC_FIELDS: &[CanonicalField] = &[
         T::Struct(&NEGATIVE_BUCKET_FIELDS),
         true,
     ),
-    F::meta(33, "summary_count", T::UInt64, true),
+    F::meta(33, "summary_count", T::Int64, true),
     F::meta(34, "summary_sum", T::Float64, true),
     F::payload(
         35,
@@ -143,7 +143,7 @@ pub static METRIC_FIELDS: &[CanonicalField] = &[
     F::sensitive(39, "exemplars", T::List(&EXEMPLAR_ELEMENT), false),
     F::meta(47, "resource_present", T::Bool, false),
     F::sensitive(48, "resource_attributes", T::Binary, false),
-    F::meta(49, "resource_dropped_attributes_count", T::UInt32, false),
+    F::meta(49, "resource_dropped_attributes_count", T::Int64, false),
     F::meta(50, "resource_schema_url", T::Utf8, false),
     F::sensitive(
         51,
@@ -155,7 +155,7 @@ pub static METRIC_FIELDS: &[CanonicalField] = &[
     F::meta(54, "scope_name", T::Utf8, false),
     F::meta(55, "scope_version", T::Utf8, false),
     F::sensitive(56, "scope_attributes", T::Binary, false),
-    F::meta(57, "scope_dropped_attributes_count", T::UInt32, false),
+    F::meta(57, "scope_dropped_attributes_count", T::Int64, false),
     F::meta(58, "scope_schema_url", T::Utf8, false),
 ];
